token fix

+++ controllers/auth.controller.js
@@ -8,23 +8,28 @@ const {
 const jwt = require("jsonwebtoken");
 
 const createRefreshToken = async (userId) => {
-  await RefreshToken.destroy({
-    where: { userId }
-  })
+  await RefreshToken.destroy({ where: { userId } });
 
-  const token = jwt.sign({ id: userId }, process.env.JWT_REFRESH_SECRET, {
-    expiresIn: "7d"
-  })
+  const generateToken = () =>
+    jwt.sign({ id: userId }, process.env.JWT_REFRESH_SECRET, { expiresIn: "7d" });
 
-  const expiresAt = new Date(Date.now() + 7 * 24 * 60 * 60 * 1000) // 7 days
+  const token = generateToken();
+  const expiresAt = new Date(Date.now() + 7 * 24 * 60 * 60 * 1000);
 
-  await RefreshToken.create({
-    token,
-    expiresAt,
-    userId
-  })
+  try {
+    await RefreshToken.create({
+      token,
+      expiresAt,
+      userId
+    });
+  } catch (err) {
+    if (err.name === "SequelizeUniqueConstraintError") {
+      return await createRefreshToken(userId);
+    }
+    throw err;
+  }
 
-  return token
+  return token;
 };
 
 exports.register = async (req, res) => {
